+++ examples/ansible_dynamic_inventory/inventory.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+cat <<EOF
+{
+    "local": {
+        "hosts": ["127.0.0.1"],
+        "vars": {
+            "ansible_connection": "local"
+        }
+    }
+}
+EOF