+++ it/server/src/test/resources/com/walmartlabs/concord/it/server/templateMerge/template/_main.js
@@ -1,0 +1,6 @@
+({
+    entryPoint: "main",
+    arguments: {
+        name: "Concord"
+    }
+})